+++ pytest_cache_assert/_check_assert/config.py
@@ -8,12 +8,12 @@ from beartype.typing import Any
 
 
 class CacheAssertContainerKeys(Enum):  # noqa: H601
-    """Enum of keys used in `cache_assert_container`."""
+    """Enum of keys used in `_cache_assert_container`."""
 
     CONFIG = '__config__'
 
 
-cache_assert_container = punq.Container()
+_cache_assert_container = punq.Container()
 """`punq` Container for injection of internal customization.
 
 Currently only supports registering custom serialization rules
@@ -24,10 +24,10 @@ Currently only supports registering custom serialization rules
 @beartype
 def register(key: CacheAssertContainerKeys, instance: Any) -> None:
     """Registers an instance for access with `retrieve(key)`."""
-    cache_assert_container.register(key, instance=instance)
+    _cache_assert_container.register(key, instance=instance)
 
 
 @beartype
 def retrieve(key: CacheAssertContainerKeys) -> Any:
     """Retrieves the value registered for the specified key."""
-    return cache_assert_container.resolve(key)
+    return _cache_assert_container.resolve(key)